perf: DataTrack page and event move to module
