+++ Sources/AnyImageKit/Capture/Util/DataTrack/Capture+AnyImageEvent.swift
@@ -1,0 +1,15 @@
+//
+//  Capture+AnyImageEvent.swift
+//  AnyImageKit
+//
+//  Created by 蒋惠 on 2021/11/12.
+//  Copyright © 2021 AnyImageProject.org. All rights reserved.
+//
+
+import Foundation
+
+extension AnyImageEvent {
+    
+    public static let capturePhoto: AnyImageEvent = "ANYIMAGEKIT_EVENT_CAPTURE_PHOTO"
+    public static let captureVideo: AnyImageEvent = "ANYIMAGEKIT_EVENT_CAPTURE_VIDEO"
+}
