+++ Sources/AnyImageKit/Capture/Util/DataTrack/Capture+AnyImagePage.swift
@@ -1,0 +1,14 @@
+//
+//  Capture+AnyImagePage.swift
+//  AnyImageKit
+//
+//  Created by 蒋惠 on 2021/11/12.
+//  Copyright © 2021 AnyImageProject.org. All rights reserved.
+//
+
+import Foundation
+
+extension AnyImagePage {
+    
+    public static let capture: AnyImagePage = "ANYIMAGEKIT_PAGE_CAPTURE"
+}
