+++ Sources/AnyImageKit/Editor/Util/DataTrack/Editor+AnyImageEvent.swift
@@ -1,0 +1,17 @@
+//
+//  Editor+AnyImageEvent.swift
+//  AnyImageKit
+//
+//  Created by 蒋惠 on 2021/11/12.
+//  Copyright © 2021 AnyImageProject.org. All rights reserved.
+//
+
+import Foundation
+
+extension AnyImageEvent {
+    
+    public static let photoPen: AnyImageEvent = "ANYIMAGEKIT_EVENT_EDITOR_PHOTO_PEN"
+    public static let photoMosaic: AnyImageEvent = "ANYIMAGEKIT_EVENT_EDITOR_PHOTO_MOSAIC"
+    public static let photoText: AnyImageEvent = "ANYIMAGEKIT_EVENT_EDITOR_PHOTO_TEXT"
+    public static let photoCrop: AnyImageEvent = "ANYIMAGEKIT_EVENT_EDITOR_PHOTO_CROP"
+}
