+++ Sources/AnyImageKit/Editor/Util/DataTrack/Editor+AnyImagePage.swift
@@ -1,0 +1,16 @@
+//
+//  Editor+AnyImagePage.swift
+//  AnyImageKit
+//
+//  Created by 蒋惠 on 2021/11/12.
+//  Copyright © 2021 AnyImageProject.org. All rights reserved.
+//
+
+import Foundation
+
+extension AnyImagePage {
+    
+    public static let photoEditor: AnyImagePage = "ANYIMAGEKIT_PAGE_EDITOR_PHOTO"
+    public static let videoEditor: AnyImagePage = "ANYIMAGEKIT_PAGE_EDITOR_VIDEO"
+    public static let textInput: AnyImagePage = "ANYIMAGEKIT_PAGE_EDITOR_TEXTINPUT"
+}
