+++ Sources/AnyImageKit/Picker/Util/DataTrack/Picker+AnyImageEvent.swift
@@ -1,0 +1,15 @@
+//
+//  Picker+AnyImageEvent.swift
+//  AnyImageKit
+//
+//  Created by 蒋惠 on 2021/11/12.
+//  Copyright © 2021 AnyImageProject.org. All rights reserved.
+//
+
+import Foundation
+
+extension AnyImageEvent {
+    
+    public static let takePhoto: AnyImageEvent = "ANYIMAGEKIT_EVENT_PICKER_TAKEPHOTO"
+    public static let takeVideo: AnyImageEvent = "ANYIMAGEKIT_EVENT_PICKER_TAKEVIDEO"
+}
