+++ Sources/AnyImageKit/Picker/Util/DataTrack/Picker+AnyImagePage.swift
@@ -1,0 +1,16 @@
+//
+//  Picker+AnyImagePage.swift
+//  AnyImageKit
+//
+//  Created by 蒋惠 on 2021/11/12.
+//  Copyright © 2021 AnyImageProject.org. All rights reserved.
+//
+
+import Foundation
+
+extension AnyImagePage {
+    
+    public static let albumPicker: AnyImagePage = "ANYIMAGEKIT_PAGE_PICKER_ALBUM"
+    public static let assetPicker: AnyImagePage = "ANYIMAGEKIT_PAGE_PICKER_ASSET"
+    public static let photoPreview: AnyImagePage = "ANYIMAGEKIT_PAGE_PICKER_PREVIEW"
+}
